Major overhaul; got simulator to generate some random events with some analysis; all parameterized
Added a SimulatorParameters class to store the existing Nodes parameter previously in CircleCalculations (although that class still bubbles up the event for the input text changing since it needs the event too); added 10 new parameters and a few readonly calculated fields for display.
Added SimulatorResult class for returning the files and events.
Added SimulatorResultAnalysis class for analyzing the data; currently just used to sum up the total buffer time which we are using as a means to measure Quality of Service (QoS).
Fleshed out a ton of stuff in the Simulation; before it only had a few hardcoded 'fake' events, but now all events are randomly generated using random distributions based on input parameters; it does routing to find files and calculates when the next event is going to fire so we don't have to wait in realtime for the simulation; it calculates when files are buffered enough to start playback (for our QoS analysis); when files are freed up from streaming they should reclaim their used bandwidth and distribute it to other nodes in order of least available bandwidth to greatest (not 100% sure that's working correcly...).
Added some various helpers and extension methods: enums/enum extensions, BinarySearch for IReadOnlyList, RandomGenerator since Math.NET Numerics Distributions need a derived Random class and won't take RandomNumberGenerator (the base class for cryptographically-secure random number generators like RNGCryptoServiceProvider).
Fixed up some CSS styling to arrange the parameter fields in a grid with a dynamic number of columns so it scales with client viewport.